fix: keep assistant-dashboard-snapshot import paths until mass rename
PR A only changes public wire identifiers; file renames land in the later mass-rename PR.
